backend/requirements.txt
asgiref==3.11.0
dj-database-url==3.0.1
Django==5.2.8
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
pillow==12.0.0
PyJWT==2.10.1
sqlparse==0.5.4
whitenoise==6.11.0

